wpt_babel/content-security-policy/object-src/object-src-url-redirect-blocked.sub.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="object-src 'self'; script-src 'self' 'unsafe-inline';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

var t = async_test("Should block the object and fire a spv");
window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
  assert_equals(e.violatedDirective, "object-src");
}));
  </script>
  <object data="/common/redirect.py?location=http://{{domains[www1]}}/content-security-policy/support/fail.png" type="image/png">
  </object>
 </body>
</html>
